scan: notify the kernel early of discovered networks
During the scan, driver receives the notification about
discovered networks spread over approximately 4-second
period. Descriptor is constructed and placed into internal
list that is then flushed in a callback function
when the scan is complete. This will cause the timestamp
associated with the network to be delayed up to 4
seconds compared to the actual time when the network is
observed. Specifically, the reported 'last seen' field in
'iw dev scan dump' will be wrong.

To fix this, call cfg80211_inform_bss as soon as
the network is observed and finish the remaining
housekeeping in the callback. This will cause the
common kernel wifi framework to generate the timestamp
with less delay than if the cfg80211_inform_bss
were called in the callback.

The remaining (and irreducable) delay is between
the hardware and the driver reciving the notification,
but this should be much less than if we waited for
the scan to complete.

Signed-off-by: Ilija Hadzic <ilijahadzic@gmail.com>